chore: Move functions to a separate module to reuse.
